feat(subagents): run async single children in process
Replaces the detached-runner single path with the in-process
continuation seam. async-execution-single.ts is deleted per spec
section 10 and its behavior moves to runs/inprocess/background-single.ts,
which admits the child, continues it in the background, and returns the
canonical child path immediately.

Refs #2188
