+++ packages/subagents/src/runs/background/async-execution.ts
@@ -2,13 +2,13 @@
  * Async execution logic for subagent tool
  */
 
+export { executeAsyncSingle } from "../inprocess/background-single.ts";
 export { executeAsyncChain } from "./async-execution-chain.ts";
 export {
 	formatAsyncStartedMessage,
 	isAsyncAvailable,
 	writeAsyncRunnerConfig,
 } from "./async-execution-common.ts";
-export { executeAsyncSingle } from "./async-execution-single.ts";
 export type {
 	AsyncChainParams,
 	AsyncExecutionContext,
