+++ test/unit/subagents-single-progress-async.test.ts
@@ -4,14 +4,9 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import type { ExtensionAPI } from "@bastani/atomic";
 import { test } from "vitest";
-import { WORKFLOW_SESSION_METADATA_ENV } from "../../packages/coding-agent/src/core/session-manager-classification.js";
 import type { AgentConfig } from "../../packages/subagents/src/agents/agent-types.js";
-import { executeAsyncSingle } from "../../packages/subagents/src/runs/background/async-execution-single.js";
+import { executeAsyncSingle } from "../../packages/subagents/src/runs/inprocess/background-single.js";
 import { ASYNC_DIR } from "../../packages/subagents/src/shared/types.js";
-
-interface CapturedRunnerConfig {
-	steps: Array<{ task: string; cwd: string }>;
-}
 
 const artifactConfig = {
 	enabled: false,
@@ -35,7 +30,11 @@ function makeAgent(): AgentConfig {
 	};
 }
 
-test("executeAsyncSingle initializes progress in isolated async storage", () => {
+function testPi(): ExtensionAPI {
+	return { events: { emit: () => {} } } as unknown as ExtensionAPI;
+}
+
+test("executeAsyncSingle initializes progress in isolated async storage", async () => {
 	const parentCwd = mkdtempSync(join(tmpdir(), "atomic-subagent-async-parent-"));
 	const childCwd = join(parentCwd, "child");
 	const artifactsDir = join(parentCwd, "disabled-artifacts");
@@ -43,15 +42,13 @@ test("executeAsyncSingle initializes progress in isolated async storage", () => 
 	const cwdProgressPath = join(childCwd, "progress.md");
 	writeFileSync(cwdProgressPath, "project sentinel");
 	const runId = `progress-${crypto.randomUUID()}`;
-	let captured: CapturedRunnerConfig | undefined;
-	let capturedEnv: Record<string, string> | undefined;
 	try {
-		const result = executeAsyncSingle(runId, {
+		const result = await executeAsyncSingle(runId, {
 			agent: "worker",
 			task: "implement the fix",
 			agentConfig: makeAgent(),
 			ctx: {
-				pi: { events: { emit: () => {} } } as unknown as ExtensionAPI,
+				pi: testPi(),
 				cwd: parentCwd,
 				currentSessionId: "parent",
 				workflowSessionMetadata: { runId: "run-1", stageId: "stage-1", stageName: "build" },
@@ -62,44 +59,34 @@ test("executeAsyncSingle initializes progress in isolated async storage", () => 
 			shareEnabled: false,
 			progress: true,
 			maxSubagentDepth: 1,
-			spawnRunner: (config, _suffix, _cwd, env) => {
-				captured = config as CapturedRunnerConfig;
-				capturedEnv = env;
-				return { pid: 1234 };
-			},
 		});
 
 		const progressPath = join(ASYNC_DIR, runId, "progress", "progress.md");
 		assert.equal(result.isError, undefined);
+		assert.equal(result.details.results[0]?.status, "continued");
+		assert.match(result.details.results[0]?.path ?? "", new RegExp(`^${runId}/`));
 		assert.equal(existsSync(join(parentCwd, "progress.md")), false, "parent cwd must not receive progress");
 		assert.equal(readFileSync(cwdProgressPath, "utf8"), "project sentinel");
 		assert.equal(existsSync(progressPath), true);
 		assert.equal(existsSync(join(artifactsDir, "progress", runId, "progress.md")), false);
 		assert.match(readFileSync(progressPath, "utf8"), /# Progress/);
-		assert.equal(captured?.steps[0]?.cwd, childCwd);
-		assert.ok((captured?.steps[0]?.task ?? "").includes(`Create and maintain progress at: ${progressPath}`));
-		assert.equal(
-			capturedEnv?.[WORKFLOW_SESSION_METADATA_ENV],
-			JSON.stringify({ runId: "run-1", stageId: "stage-1", stageName: "build" }),
-		);
 	} finally {
 		rmSync(join(ASYNC_DIR, runId), { recursive: true, force: true });
 		rmSync(parentCwd, { recursive: true, force: true });
 	}
 });
 
-test("executeAsyncSingle prefers run-scoped artifact storage", () => {
+test("executeAsyncSingle prefers run-scoped artifact storage", async () => {
 	const parentCwd = mkdtempSync(join(tmpdir(), "atomic-subagent-async-artifacts-"));
 	const artifactsDir = join(parentCwd, "artifacts");
 	const runId = `progress-${crypto.randomUUID()}`;
-	let captured: CapturedRunnerConfig | undefined;
 	try {
-		const result = executeAsyncSingle(runId, {
+		const result = await executeAsyncSingle(runId, {
 			agent: "worker",
 			task: "implement the fix",
 			agentConfig: makeAgent(),
 			ctx: {
-				pi: { events: { emit: () => {} } } as unknown as ExtensionAPI,
+				pi: testPi(),
 				cwd: parentCwd,
 				currentSessionId: "parent",
 			},
@@ -108,16 +95,12 @@ test("executeAsyncSingle prefers run-scoped artifact storage", () => {
 			shareEnabled: false,
 			progress: true,
 			maxSubagentDepth: 1,
-			spawnRunner: (config) => {
-				captured = config as CapturedRunnerConfig;
-				return { pid: 1234 };
-			},
 		});
 
 		const progressPath = join(artifactsDir, "progress", runId, "progress.md");
 		assert.equal(result.isError, undefined);
+		assert.equal(result.details.results[0]?.status, "continued");
 		assert.equal(existsSync(progressPath), true);
-		assert.ok((captured?.steps[0]?.task ?? "").includes(`Create and maintain progress at: ${progressPath}`));
 		assert.equal(existsSync(join(parentCwd, "progress.md")), false);
 	} finally {
 		rmSync(join(ASYNC_DIR, runId), { recursive: true, force: true });
